Split UART into hardware independent and hardware specific parts
