+++ src/system/bus/uart/uart.c
@@ -1,29 +1,22 @@
 /**
  * @file uart.c
- * @brief UART (Universal Asynchronous Receiver/Transmitter) module implementation
+ * @brief Hardware-independent UART (Universal Asynchronous Receiver/Transmitter) implementation
  *
- * This module provides functions and routines for configuring and handling
- * UART communication. It includes initialization, data transmission, and
- * reception functionalities for serial communication.
+ * This module provides the hardware-independent layer of the UART driver.
+ * It manages circular buffers for transmission and reception, handles callbacks,
+ * and provides the public API exposed through uart.h.
  *
- * Implementation Details:
- * - Uses USART3 on STM32H5 (PD8=TX, PD9=RX)
- * - Configured for 115200 baud, 8N1 format
- * - Interrupt-driven with circular buffers (256 bytes each for TX/RX)
- * - Direct RX interrupt handling for optimal performance
- * - DMA-based reception with idle line detection
- * - DMA-based transmission for optimal performance
- * - Supports configurable RX callbacks for protocol implementations
- * - NVIC priority set to 3 for UART interrupts
- *
- * Buffer Management:
- * - RX buffer: Filled by DMA, emptied by UART_read()
- * - TX buffer: Filled by UART_write(), emptied by DMA
- * - Buffer full conditions return errors rather than blocking
+ * Features:
+ * - Non-blocking read/write operations with circular buffers
+ * - Configurable RX callback for protocol implementations
+ * - Buffer status inquiry functions
  * - Circular buffer implementation with automatic wrap-around
  *
+ * This implementation relies on hardware-specific functions defined in
+ * src/system/h563xx/uart.c (or equivalent for other platforms).
+ *
  * @author Alexander Bessman
- * @date 2025-06-27
+ * @date 2025-06-28
  */
 
 #include <stdbool.h>
@@ -31,29 +24,19 @@
 #include <stdint.h>
 #include <string.h>
 
-#include "stm32h5xx_hal.h"
-
 #include "bus_common.h"
 #include "uart.h"
+#include "uart_internal.h"
 
-
+/* Buffer sizes - must be powers of 2 */
 #define UART_RX_BUFFER_SIZE 256
 #define UART_TX_BUFFER_SIZE 256
-
-
-static UART_HandleTypeDef huart = { 0 };
-static DMA_HandleTypeDef hdma_usart3_tx = { 0 };
-static DMA_HandleTypeDef hdma_usart3_rx = { 0 };
 
 /* Buffers and their management structures */
 static uint8_t rx_buffer_data[UART_RX_BUFFER_SIZE];
 static uint8_t tx_buffer_data[UART_TX_BUFFER_SIZE];
 static circular_buffer_t rx_buffer;
 static circular_buffer_t tx_buffer;
-
-/* Interrupt state */
-static bool volatile tx_in_progress = false;
-static uint32_t volatile tx_dma_size = 0;
 
 /* RX DMA state */
 static uint32_t volatile rx_dma_head = 0;
@@ -62,6 +45,28 @@ static uint32_t volatile rx_dma_head = 0;
 static uart_rx_callback_t rx_callback = NULL;
 static uint32_t rx_threshold = 0;
 
+/* Forward declarations of hardware-specific functions */
+void UART_set_dma_buffer(uint8_t *buffer, uint32_t size);
+void UART_start_dma_tx(uint8_t *buffer, uint32_t size);
+uint32_t UART_get_dma_position(void);
+bool UART_hw_tx_busy(void);
+
+
+/**
+ * @brief Buffer initialization called from hardware-specific init
+ *
+ * @param buffer Buffer to use for RX DMA (ignored, we use our own)
+ * @param size Size of the buffer (ignored)
+ */
+void UART_buffer_init(void)
+{
+    /* Initialize circular buffers */
+    circular_buffer_init(&rx_buffer, rx_buffer_data, UART_RX_BUFFER_SIZE);
+    circular_buffer_init(&tx_buffer, tx_buffer_data, UART_TX_BUFFER_SIZE);
+
+    /* Set the DMA buffer in the hardware layer */
+    UART_set_dma_buffer(rx_buffer_data, UART_RX_BUFFER_SIZE);
+}
 
 /**
  * @brief Get number of bytes available in TX buffer.
@@ -77,7 +82,7 @@ static uint32_t tx_buffer_available(void)
 static uint32_t rx_buffer_available(void)
 {
     /* Get current DMA write position */
-    uint32_t dma_pos = UART_RX_BUFFER_SIZE - __HAL_DMA_GET_COUNTER(&hdma_usart3_rx);
+    uint32_t dma_pos = UART_get_dma_position();
     uint32_t dma_head = dma_pos;
 
     /* Update our cached head position */
@@ -95,7 +100,7 @@ static uint32_t rx_buffer_available(void)
  */
 static void start_transmission(void)
 {
-    if (tx_in_progress || circular_buffer_is_empty(&tx_buffer)) {
+    if (UART_hw_tx_busy() || circular_buffer_is_empty(&tx_buffer)) {
         return;
     }
 
@@ -119,42 +124,61 @@ static void start_transmission(void)
     }
 
     if (contiguous_bytes > 0) {
-        tx_in_progress = true;
-        tx_dma_size = contiguous_bytes;
-        HAL_UART_Transmit_DMA(&huart, &tx_buffer.buffer[tx_buffer.tail], contiguous_bytes);
+        /* Start hardware TX DMA */
+        UART_start_dma_tx(&tx_buffer.buffer[tx_buffer.tail], contiguous_bytes);
     }
 }
 
 /**
- * @brief Initialize the UART peripheral.
+ * @brief Callback invoked by hardware layer when TX is complete
  *
- * This function configures the UART hardware, including baud rate, data bits,
- * stop bits, and parity, to prepare it for serial communication.
+ * @param bytes_transferred Number of bytes transferred
  */
-void UART_init(void)
+void UART_tx_complete_callback(uint32_t bytes_transferred)
 {
-    /* Initialize circular buffers */
-    circular_buffer_init(&rx_buffer, rx_buffer_data, UART_RX_BUFFER_SIZE);
-    circular_buffer_init(&tx_buffer, tx_buffer_data, UART_TX_BUFFER_SIZE);
+    /* Update TX buffer tail with the number of bytes that were sent */
+    tx_buffer.tail = (tx_buffer.tail + bytes_transferred) % tx_buffer.size;
 
-    huart.Instance = USART3;
-    huart.Init.BaudRate = 115200;
-    huart.Init.WordLength = UART_WORDLENGTH_8B;
-    huart.Init.StopBits = UART_STOPBITS_1;
-    huart.Init.Parity = UART_PARITY_NONE;
-    huart.Init.Mode = UART_MODE_TX_RX;
-    HAL_UART_Init(&huart);
+    /* Try to start another transmission if there's more data */
+    start_transmission();
+}
 
-    /* Start DMA reception
-     * by restarting transfers in HAL_UART_RxCpltCallback */
-    HAL_UART_Receive_DMA(&huart, rx_buffer_data, UART_RX_BUFFER_SIZE);
+/**
+ * @brief Check if RX callback condition is met and call if needed
+ */
+static bool check_rx_callback(void)
+{
+    if (rx_callback && rx_buffer_available() >= rx_threshold) {
+        rx_callback(rx_buffer_available());
+        return true;
+    }
+    return false;
+}
 
-    /* Enable UART idle line interrupt for packet detection */
-    __HAL_UART_ENABLE_IT(&huart, UART_IT_IDLE);
+/**
+ * @brief Callback invoked by hardware layer when RX DMA buffer is full
+ *
+ * @param buffer_size Size of the buffer
+ */
+void UART_rx_complete_callback(uint32_t buffer_size)
+{
+    /* In circular DMA mode, just check for callbacks */
+    check_rx_callback();
+}
 
-    /* Clear any pending flags */
-    __HAL_UART_CLEAR_FLAG(&huart, UART_FLAG_IDLE);
-    __HAL_UART_CLEAR_FLAG(&huart, UART_FLAG_ORE);
+/**
+ * @brief Callback invoked by hardware layer on UART idle line detection
+ *
+ * @param dma_pos Current DMA position
+ */
+void UART_idle_callback(uint32_t dma_pos)
+{
+    /* Update our cached DMA position */
+    rx_dma_head = dma_pos;
+    rx_buffer.head = dma_pos;
+
+    /* Check for callbacks */
+    check_rx_callback();
 }
 
 /**
@@ -252,148 +276,7 @@ uint32_t UART_tx_free_space(void)
  */
 bool UART_tx_busy(void)
 {
-    return tx_in_progress;
-}
-
-/**
- * @brief  Initialize the UART MSP (MCU Support Package).
- *
- * This function configures the hardware resources used for the UART
- * peripheral, including GPIO pins and clocks. It is called by
- * HAL_UART_Init().
- *
- * @param huart UART handle pointer.
- *
- * @return None
- */
-void HAL_UART_MspInit(UART_HandleTypeDef *const huart_ptr)
-{
-    __HAL_RCC_GPIOD_CLK_ENABLE();
-    __HAL_RCC_USART3_CLK_ENABLE();
-    __HAL_RCC_GPDMA1_CLK_ENABLE();
-
-    GPIO_InitTypeDef uart_gpio_init = {
-        .Pin = GPIO_PIN_8 | GPIO_PIN_9,
-        .Mode = GPIO_MODE_AF_PP,
-        .Pull = GPIO_NOPULL,
-        .Speed = GPIO_SPEED_FREQ_HIGH,
-        .Alternate = GPIO_AF7_USART3
-    };
-    HAL_GPIO_Init(GPIOD, &uart_gpio_init);
-
-    /* Configure DMA for TX */
-    hdma_usart3_tx.Instance = GPDMA1_Channel0;
-    hdma_usart3_tx.Init.Request = GPDMA1_REQUEST_USART3_TX;
-    hdma_usart3_tx.Init.BlkHWRequest = DMA_BREQ_SINGLE_BURST;
-    hdma_usart3_tx.Init.Direction = DMA_MEMORY_TO_PERIPH;
-    hdma_usart3_tx.Init.SrcInc = DMA_SINC_INCREMENTED;
-    hdma_usart3_tx.Init.DestInc = DMA_DINC_FIXED;
-    hdma_usart3_tx.Init.SrcDataWidth = DMA_SRC_DATAWIDTH_BYTE;
-    hdma_usart3_tx.Init.DestDataWidth = DMA_DEST_DATAWIDTH_BYTE;
-    hdma_usart3_tx.Init.Priority = DMA_LOW_PRIORITY_LOW_WEIGHT;
-    hdma_usart3_tx.Init.SrcBurstLength = 1;
-    hdma_usart3_tx.Init.DestBurstLength = 1;
-    hdma_usart3_tx.Init.TransferAllocatedPort = (
-        DMA_SRC_ALLOCATED_PORT0 | DMA_DEST_ALLOCATED_PORT1
-    );
-    hdma_usart3_tx.Init.TransferEventMode = DMA_TCEM_BLOCK_TRANSFER;
-    hdma_usart3_tx.Init.Mode = DMA_NORMAL;
-
-    HAL_DMA_Init(&hdma_usart3_tx);
-
-    /* Configure DMA for RX */
-    hdma_usart3_rx.Instance = GPDMA1_Channel1;
-    hdma_usart3_rx.Init.Request = GPDMA1_REQUEST_USART3_RX;
-    hdma_usart3_rx.Init.BlkHWRequest = DMA_BREQ_SINGLE_BURST;
-    hdma_usart3_rx.Init.Direction = DMA_PERIPH_TO_MEMORY;
-    hdma_usart3_rx.Init.SrcInc = DMA_SINC_FIXED;
-    hdma_usart3_rx.Init.DestInc = DMA_DINC_INCREMENTED;
-    hdma_usart3_rx.Init.SrcDataWidth = DMA_SRC_DATAWIDTH_BYTE;
-    hdma_usart3_rx.Init.DestDataWidth = DMA_DEST_DATAWIDTH_BYTE;
-    hdma_usart3_rx.Init.Priority = DMA_LOW_PRIORITY_LOW_WEIGHT;
-    hdma_usart3_rx.Init.SrcBurstLength = 1;
-    hdma_usart3_rx.Init.DestBurstLength = 1;
-    hdma_usart3_rx.Init.TransferAllocatedPort = (
-        DMA_SRC_ALLOCATED_PORT0 | DMA_DEST_ALLOCATED_PORT1
-    );
-    hdma_usart3_rx.Init.TransferEventMode = DMA_TCEM_BLOCK_TRANSFER;
-    /* Use normal mode, manage circular manually */
-    hdma_usart3_rx.Init.Mode = DMA_NORMAL;
-
-    HAL_DMA_Init(&hdma_usart3_rx);
-
-    /* Link DMA to UART */
-    __HAL_LINKDMA(huart_ptr, hdmatx, hdma_usart3_tx);
-    __HAL_LINKDMA(huart_ptr, hdmarx, hdma_usart3_rx);
-
-    /* Configure NVIC for UART and DMA interrupts */
-    HAL_NVIC_SetPriority(USART3_IRQn, 3, 0);
-    HAL_NVIC_EnableIRQ(USART3_IRQn);
-
-    HAL_NVIC_SetPriority(GPDMA1_Channel0_IRQn, 3, 1);
-    HAL_NVIC_EnableIRQ(GPDMA1_Channel0_IRQn);
-
-    HAL_NVIC_SetPriority(GPDMA1_Channel1_IRQn, 3, 1);
-    HAL_NVIC_EnableIRQ(GPDMA1_Channel1_IRQn);
-}
-
-/**
- * @brief UART interrupt handler.
- *
- * This function handles RX interrupts directly for better performance,
- * while delegating TX and other interrupts to the HAL handler.
- */
-void USART3_IRQHandler(void)
-{
-    /* Handle idle line interrupt for packet detection */
-    if (__HAL_UART_GET_FLAG(&huart, UART_FLAG_IDLE)
-        && __HAL_UART_GET_IT_SOURCE(&huart, UART_IT_IDLE)
-    ) {
-        __HAL_UART_CLEAR_IDLEFLAG(&huart);
-
-        /* Check for RX callback */
-        if (rx_callback && rx_buffer_available() >= rx_threshold) {
-            rx_callback(rx_buffer_available());
-        }
-    }
-
-    /* Handle other interrupts through HAL */
-    HAL_UART_IRQHandler(&huart);
-}
-
-/**
- * @brief TX Complete callback from HAL.
- *
- * This function is called when a DMA transmission is complete.
- */
-void HAL_UART_TxCpltCallback(UART_HandleTypeDef *const huart_ptr)
-{
-    if (huart_ptr->Instance == USART3) {
-        /* Update tail position to reflect transmitted data */
-        tx_buffer.tail = (tx_buffer.tail + tx_dma_size) % tx_buffer.size;
-        tx_in_progress = false;
-
-        /* Check if more data needs to be transmitted */
-        start_transmission();
-    }
-}
-
-/**
- * @brief RX Complete callback from HAL.
- *
- * This function is called when DMA RX is complete (i.e., buffer full).
- */
-void HAL_UART_RxCpltCallback(UART_HandleTypeDef *const huart_ptr)
-{
-    if (huart_ptr->Instance == USART3) {
-        /* Restart DMA reception */
-        HAL_UART_Receive_DMA(&huart, rx_buffer_data, UART_RX_BUFFER_SIZE);
-
-        /* Check for RX callback */
-        if (rx_callback && rx_buffer_available() >= rx_threshold) {
-            rx_callback(rx_buffer_available());
-        }
-    }
+    return UART_hw_tx_busy();
 }
 
 /**
@@ -413,20 +296,4 @@ void UART_set_rx_callback(
     if (rx_callback && rx_buffer_available() >= rx_threshold) {
         rx_callback(rx_buffer_available());
     }
-}
-
-/**
- * @brief DMA TX interrupt handler.
- */
-void GPDMA1_Channel0_IRQHandler(void)
-{
-    HAL_DMA_IRQHandler(&hdma_usart3_tx);
-}
-
-/**
- * @brief DMA RX interrupt handler.
- */
-void GPDMA1_Channel1_IRQHandler(void)
-{
-    HAL_DMA_IRQHandler(&hdma_usart3_rx);
 }
